Use the StackView properly
Previously we were creating all the views and just marking them as
invisible. Now each view is its own Component which is created and
destroyed when it is pushed into the StackView.

This does however result in it being slightly slower to switch from the
different views as they need to be created and filled, but that's how it
should be. We will also consume less memory this way.