Add a special ImageBrowser in the mobile view
In the mobile view we want to be able to flick through images, the only
way I could think of doing that is via a ListView with the delegate
consuming the entire screen space.

Also, on a mobile we really do not need the complex ImageViewer which
shows a toolbar and what not.